pyproject.toml
[project]
name = "vlm-semantic-spatial-perception"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = "==3.12.3"
dependencies = [
    "anthropic>=0.25.0",
    "black>=23.0.0",
    "flake8>=6.0.0",
    "google-genai>=0.1.0",
    "google-generativeai>=0.3.0",
    "loguru>=0.7.0",
    "matplotlib>=3.7.0",
    "mypy>=1.5.0",
    "numpy>=1.24.0",
    "open3d>=0.18.0",
    "sam2>=1.0.0",
    "supervision>=0.22.0",
    "ram @ git+https://github.com/xinyu1205/recognize-anything.git",
    "openai>=1.0.0",
    "opencv-python>=4.8.0",
    "pddl>=0.3.0",
    "pillow>=10.0.0",
    "plotly>=5.18.0",
    "pyrealsense2>=2.55.0",
    "pytest>=7.4.0",
    "pytest-asyncio>=0.23.0",
    "pytest-cov>=4.1.0",
    "python-dotenv>=1.0.0",
    "pyyaml>=6.0",
    "scikit-learn>=1.3.0",
    "scipy>=1.11.0",
    "textual==0.56.4",
    "tqdm>=4.66.0",
    "torch>=2.9.0",
    "transformers>=4.45.0",
    "accelerate>=0.34.0",
    "xarm-python-sdk>=1.17.0",
    "pyperplan>=2.1",
    "pybullet>=3.2.7",
    "timm>=1.0.25",
    "fairscale>=0.4.13",
    "bitsandbytes>=0.49.2",
]
torch = [
  { index = "pytorch-cu129", marker = "sys_platform == 'linux' or sys_platform == 'win32'" },
]
[[tool.uv.index]]
name = "pytorch-cu129"
url = "https://download.pytorch.org/whl/cu129"
explicit = true
